Move container to a container folder
